moved windowing and rendering into a shared "rendering" folder, added more license info to the windowing interface, added more getters to the windowing interface for things like height and title
